Add support for quadratic programs
This adds support for solving quadratic programs on the same footing as
(mixed integer) linear programs. HiGHS itself supports these programs
out of the box, so the main change necessary here is to ensure that all
outputs are parsed as expected.

One complicating factor is that the collection of populated columns is
different: Just like infeasible problems lack some of the columns, for
quadratic programs, the "Status" column is not populated. We take care
of this by following the same pattern as for the other cases, simply
filtering the header before parsing the result; to do so, we keep track
of whether we are looking at a quadratic program and as a bonus, we add
a property called `IsQuadratic` to the output.

A couple of tests are added to ensure that this works as expected. As it
turns out, the version of HiGHS added as a submodule produces incorrect
results for one of the test cases, by producing non-error output for a
problem whose Hessian isn't positive semi-definite. To get around this
problem, we simply update the HiGHS submodule to the most recent
version.

Another change since the previously bundled version of HiGHS is that the
output now contains the value of the objective as well. While this seems
like something that could be useful to include in the JavaScript output
object as well, here we just opt to remove it before parsing, leaving
the parsing of the objective value as an exercise for the future.