Parse headers by looking at data lines instead of looking at problem type
Before this, the headers available for the columns/rows were determined
by looking at the type of problem under consideration. This approach
wouldn't quite cut it in the case of quadratic programs as there was no
way to tell that the problem was a QP without looking at the input. Here
we instead determine the relevant headers by looking at the row
immediately below the header line.