Avoid treating loadModel warnings as errors
The C API function `readModel` can return one of three values, cf.

```c
enum class HighsStatus { kError = -1, kOk = 0, kWarning = 1 };
```

and in the corresponding piece of JavaScript, `assert_ok` would make the
requirement that result is 0. Here, we change that to require only that
the result is not -1, thus allowing `solve` to continue in case the
loading of the model causes a warning.

Note also that it is possible for `readModel` to error completely (as is
the case for the test `test_invalid_model`), so we make sure to
explicitly check for that as well.

It could be useful enough to actually provide the warning message to the
user somehow, but figuring out what the warning message is might be some
work; it is not part of the pretty output, for example.