Remove "IsLinear" and "IsQuadratic" from solution object
The information is rather superfluous, as, on one hand, the user will
likely be aware which model they are solving, and on the other, since we
can include the types of all variables now, the information on whether
a model has integer variables can be determined from that.

Indeed, we update all tests to contain the type information as well.